# functional/src/main/java/com/rjtoursim/api/controller/V1Api.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.13.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.rjtoursim.api.controller;

import com.rjtoursim.api.model.CalculateAchievementResponse;
import com.rjtoursim.api.model.CategoryResponse;
import com.rjtoursim.api.model.CreatePostRequest;
import com.rjtoursim.api.model.FetchLikesResponse;
import com.rjtoursim.api.model.FetchPostsResponse;
import com.rjtoursim.api.model.GetLikeResponse;
import com.rjtoursim.api.model.LikeRequest;
import com.rjtoursim.api.model.UserCredentials;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-03-18T23:02:07.666399Z[Europe/London]", comments = "Generator version: 7.13.0")
@Validated
@Tag(name = "v1", description = "the v1 API")
public interface V1Api {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /v1/content/add-like : Add a like to a post with a given user ID and post ID.
     *
     * @param likeRequest  (optional)
     * @return OK - Like added successfully (status code 200)
     *         or Not Found - User or post with the given ID does not exist (status code 404)
     */
    @Operation(
        operationId = "addLike",
        summary = "Add a like to a post with a given user ID and post ID.",
        responses = {
            @ApiResponse(responseCode = "200", description = "OK - Like added successfully"),
            @ApiResponse(responseCode = "404", description = "Not Found - User or post with the given ID does not exist")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/v1/content/add-like",
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Void> _addLike(
        @Parameter(name = "LikeRequest", description = "") @Valid @RequestBody(required = false) LikeRequest likeRequest
    ) {
        return addLike(likeRequest);
    }

    // Override this method
    default  ResponseEntity<Void> addLike(LikeRequest likeRequest) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /v1/authentication/authenticate-user : For a given username and password, authenticate them.
     *
     * @param userCredentials  (optional)
     * @return Authenitcation successfull (status code 200)
     *         or Unauthorized - Invalid username or password (status code 401)
     *         or Internal Server Error - Something went wrong on the server side (status code 500)
     */
    @Operation(
        operationId = "authenticateUser",
        summary = "For a given username and password, authenticate them.",
        responses = {
            @ApiResponse(responseCode = "200", description = "Authenitcation successfull"),
            @ApiResponse(responseCode = "401", description = "Unauthorized - Invalid username or password"),
            @ApiResponse(responseCode = "500", description = "Internal Server Error - Something went wrong on the server side")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/v1/authentication/authenticate-user",
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Void> _authenticateUser(
        @Parameter(name = "UserCredentials", description = "") @Valid @RequestBody(required = false) UserCredentials userCredentials
    ) {
        return authenticateUser(userCredentials);
    }

    // Override this method
    default  ResponseEntity<Void> authenticateUser(UserCredentials userCredentials) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /v1/account/calculate-achievement/{username} : Calculate the achievements for a given user ID.
     *
     * @param username  (required)
     * @return OK (status code 200)
     *         or Not Found - User with the given ID does not exist (status code 404)
     */
    @Operation(
        operationId = "calculateAchievement",
        summary = "Calculate the achievements for a given user ID.",
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CalculateAchievementResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Not Found - User with the given ID does not exist")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/v1/account/calculate-achievement/{username}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<CalculateAchievementResponse> _calculateAchievement(
        @Parameter(name = "username", description = "", required = true, in = ParameterIn.PATH) @PathVariable("username") String username
    ) {
        return calculateAchievement(username);
    }

    // Override this method
    default  ResponseEntity<CalculateAchievementResponse> calculateAchievement(String username) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"FiveLikesAchievement\" : true }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /v1/content/create-post : Create a post with a given title and content.
     *
     * @param createPostRequest  (optional)
     * @return Created (status code 201)
     */
    @Operation(
        operationId = "createPost",
        summary = "Create a post with a given title and content.",
        responses = {
            @ApiResponse(responseCode = "201", description = "Created")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/v1/content/create-post",
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Void> _createPost(
        @Parameter(name = "CreatePostRequest", description = "") @Valid @RequestBody(required = false) CreatePostRequest createPostRequest
    ) {
        return createPost(createPostRequest);
    }

    // Override this method
    default  ResponseEntity<Void> createPost(CreatePostRequest createPostRequest) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /v1/account/create-user : Create a user account with a given username and password.
     *
     * @param userCredentials  (optional)
     * @return Created (status code 201)
     *         or Conflict - Username already exists (status code 409)
     */
    @Operation(
        operationId = "createUser",
        summary = "Create a user account with a given username and password.",
        responses = {
            @ApiResponse(responseCode = "201", description = "Created"),
            @ApiResponse(responseCode = "409", description = "Conflict - Username already exists")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/v1/account/create-user",
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Void> _createUser(
        @Parameter(name = "UserCredentials", description = "") @Valid @RequestBody(required = false) UserCredentials userCredentials
    ) {
        return createUser(userCredentials);
    }

    // Override this method
    default  ResponseEntity<Void> createUser(UserCredentials userCredentials) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /v1/content/delete-comment/{commentId} : Delete a comment with a given comment ID.
     *
     * @param commentId  (required)
     * @return OK - Comment deleted successfully (status code 200)
     *         or Not Found - Comment with the given ID does not exist (status code 404)
     */
    @Operation(
        operationId = "deleteComment",
        summary = "Delete a comment with a given comment ID.",
        responses = {
            @ApiResponse(responseCode = "200", description = "OK - Comment deleted successfully"),
            @ApiResponse(responseCode = "404", description = "Not Found - Comment with the given ID does not exist")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/v1/content/delete-comment/{commentId}"
    )
    
    default ResponseEntity<Void> _deleteComment(
        @Parameter(name = "commentId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("commentId") Integer commentId
    ) {
        return deleteComment(commentId);
    }

    // Override this method
    default  ResponseEntity<Void> deleteComment(Integer commentId) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /v1/content/delete-post/{postId} : Delete a post with a given post ID.
     *
     * @param postId  (required)
     * @return OK - Post deleted successfully (status code 200)
     *         or Not Found - Post with the given ID does not exist (status code 404)
     */
    @Operation(
        operationId = "deletePost",
        summary = "Delete a post with a given post ID.",
        responses = {
            @ApiResponse(responseCode = "200", description = "OK - Post deleted successfully"),
            @ApiResponse(responseCode = "404", description = "Not Found - Post with the given ID does not exist")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/v1/content/delete-post/{postId}"
    )
    
    default ResponseEntity<Void> _deletePost(
        @Parameter(name = "postId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("postId") Integer postId
    ) {
        return deletePost(postId);
    }

    // Override this method
    default  ResponseEntity<Void> deletePost(Integer postId) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /v1/content/fetch-likes/{postId} : Fetch the number of likes for a given post ID.
     *
     * @param postId  (required)
     * @return OK - Number of likes retrieved successfully (status code 200)
     *         or Not Found - Post with the given ID does not exist (status code 404)
     */
    @Operation(
        operationId = "fetchLikes",
        summary = "Fetch the number of likes for a given post ID.",
        responses = {
            @ApiResponse(responseCode = "200", description = "OK - Number of likes retrieved successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = FetchLikesResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Not Found - Post with the given ID does not exist")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/v1/content/fetch-likes/{postId}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<FetchLikesResponse> _fetchLikes(
        @Parameter(name = "postId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("postId") Integer postId
    ) {
        return fetchLikes(postId);
    }

    // Override this method
    default  ResponseEntity<FetchLikesResponse> fetchLikes(Integer postId) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"likeCount\" : 0 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /v1/content/fetch-posts : Fetch all posts.
     *
     * @return OK (status code 200)
     *         or Not Found - No posts found (status code 404)
     */
    @Operation(
        operationId = "fetchPosts",
        summary = "Fetch all posts.",
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = FetchPostsResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Not Found - No posts found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/v1/content/fetch-posts",
        produces = { "application/json" }
    )
    
    default ResponseEntity<FetchPostsResponse> _fetchPosts(
        
    ) {
        return fetchPosts();
    }

    // Override this method
    default  ResponseEntity<FetchPostsResponse> fetchPosts() {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"posts\" : [ { \"address\" : \"address\", \"description\" : \"description\", \"id\" : 0, \"title\" : \"title\", \"datePosted\" : \"2000-01-23T04:56:07.000+00:00\", \"categoryId\" : 6, \"username\" : \"username\" }, { \"address\" : \"address\", \"description\" : \"description\", \"id\" : 0, \"title\" : \"title\", \"datePosted\" : \"2000-01-23T04:56:07.000+00:00\", \"categoryId\" : 6, \"username\" : \"username\" } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /v1/content/fetch-posts-by-category/{catId} : Fetch posts by a given category id.
     *
     * @param catId  (required)
     * @return OK (status code 200)
     *         or Not Found - No posts found (status code 404)
     */
    @Operation(
        operationId = "fetchPostsByCategory",
        summary = "Fetch posts by a given category id.",
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = FetchPostsResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Not Found - No posts found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/v1/content/fetch-posts-by-category/{catId}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<FetchPostsResponse> _fetchPostsByCategory(
        @Parameter(name = "catId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("catId") Integer catId
    ) {
        return fetchPostsByCategory(catId);
    }

    // Override this method
    default  ResponseEntity<FetchPostsResponse> fetchPostsByCategory(Integer catId) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"posts\" : [ { \"address\" : \"address\", \"description\" : \"description\", \"id\" : 0, \"title\" : \"title\", \"datePosted\" : \"2000-01-23T04:56:07.000+00:00\", \"categoryId\" : 6, \"username\" : \"username\" }, { \"address\" : \"address\", \"description\" : \"description\", \"id\" : 0, \"title\" : \"title\", \"datePosted\" : \"2000-01-23T04:56:07.000+00:00\", \"categoryId\" : 6, \"username\" : \"username\" } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /v1/content/get-category/{catId} : Get the associated category name for an id.
     *
     * @param catId  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "getCategory",
        summary = "Get the associated category name for an id.",
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CategoryResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/v1/content/get-category/{catId}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<CategoryResponse> _getCategory(
        @Parameter(name = "catId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("catId") Integer catId
    ) {
        return getCategory(catId);
    }

    // Override this method
    default  ResponseEntity<CategoryResponse> getCategory(Integer catId) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"categoryName\" : \"categoryName\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /v1/content/get-like : Get the like status for a given user ID and post ID.
     *
     * @param likeRequest  (optional)
     * @return OK - Like status retrieved successfully (status code 200)
     *         or Not Found - User or post with the given ID does not exist (status code 404)
     */
    @Operation(
        operationId = "getLike",
        summary = "Get the like status for a given user ID and post ID.",
        responses = {
            @ApiResponse(responseCode = "200", description = "OK - Like status retrieved successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = GetLikeResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Not Found - User or post with the given ID does not exist")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/v1/content/get-like",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<GetLikeResponse> _getLike(
        @Parameter(name = "LikeRequest", description = "") @Valid @RequestBody(required = false) LikeRequest likeRequest
    ) {
        return getLike(likeRequest);
    }

    // Override this method
    default  ResponseEntity<GetLikeResponse> getLike(LikeRequest likeRequest) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"likeStatus\" : true }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /v1/authentication : Ping authentication controller to see if it is reachable
     *
     * @return Pinged successfully (status code 200)
     */
    @Operation(
        operationId = "pingAuth",
        summary = "Ping authentication controller to see if it is reachable",
        responses = {
            @ApiResponse(responseCode = "200", description = "Pinged successfully")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/v1/authentication"
    )
    
    default ResponseEntity<Void> _pingAuth(
        
    ) {
        return pingAuth();
    }

    // Override this method
    default  ResponseEntity<Void> pingAuth() {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}