# presentation/presentation-layer/api-spec.yaml
openapi: 3.0.0
info:
  title: "API"
  version: "1.0.0"
paths:
  /v1/authentication:
    get:
      operationId: ping-auth
      summary: Ping authentication controller to see if it is reachable
      responses:
        '200':
          description: Pinged successfully
  '/v1/authentication/authenticate-user':
    post:
      operationId: authenticate-user
      summary: For a given username and password, authenticate them.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCredentials'
      responses:
        '200':
          description: Authenitcation successfull
        '401':
          description: Unauthorized - Invalid username or password
        '500':
          description: Internal Server Error - Something went wrong on the server side
  '/v1/account/create-user':
    post:
      operationId: create-user
      summary: Create a user account with a given username and password.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCredentials'
      responses:
        '201':
          description: Created
        '409':
          description: Conflict - Username already exists
  '/v1/account/calculate-achievement/{username}':
    get:
      operationId: calculate-achievement
      summary: Calculate the achievements for a given user ID.
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                title: CalculateAchievementResponse
                type: object
                properties:
                  FiveLikesAchievement:
                    type: boolean
        '404':
          description: Not Found - User with the given ID does not exist
  '/v1/content/create-post':
    post:
      operationId: create-post
      summary: Create a post with a given title and content.
      requestBody:
        content:
          application/json:
            schema:
              title: CreatePostRequest
              type: object
              properties:
                username:
                  type: string
                categoryId:
                  type: integer
                title:
                  type: string
                description:
                  type: string
                address:
                  type: string
                datePosted:
                  type: string
                  format: date-time
      responses:
        '201':
          description: Created
  '/v1/content/get-category/{catId}':
    get:
      operationId: get-category
      summary: Get the associated category name for an id.
      parameters:
        - name: catId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                title: CategoryResponse
                type: object
                properties:
                  categoryName:
                    type: string
  '/v1/content/fetch-posts':
    get:
      operationId: fetch-posts
      summary: Fetch all posts.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FetchPostsResponse'
        '404':
          description: Not Found - No posts found
  '/v1/content/fetch-posts-by-category/{catId}':
    get:
      operationId: fetch-posts-by-category
      summary: Fetch posts by a given category id.
      parameters:
        - name: catId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FetchPostsResponse'
        '404':
          description: Not Found - No posts found
  '/v1/content/delete-post/{postId}':
    delete:
      operationId: delete-post
      summary: Delete a post with a given post ID.
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK - Post deleted successfully
        '404':
          description: Not Found - Post with the given ID does not exist
  '/v1/content/delete-comment/{commentId}':
    delete:
      operationId: delete-comment
      summary: Delete a comment with a given comment ID.
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK - Comment deleted successfully
        '404':
          description: Not Found - Comment with the given ID does not exist
  '/v1/content/add-like':
    post:
      operationId: add-like
      summary: Add a like to a post with a given user ID and post ID.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LikeRequest'
      responses:
        '200':
          description: OK - Like added successfully
        '404':
          description: Not Found - User or post with the given ID does not exist
  '/v1/content/get-like':
    post:
      operationId: get-like
      summary: Get the like status for a given user ID and post ID.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LikeRequest'
      responses:
        '200':
          description: OK - Like status retrieved successfully
          content:
            application/json:
              schema:
                title: GetLikeResponse
                type: object
                properties:
                  likeStatus:
                    type: boolean
        '404':
          description: Not Found - User or post with the given ID does not exist
  '/v1/content/fetch-likes/{postId}':
    get:
      operationId: fetch-likes
      summary: Fetch the number of likes for a given post ID.
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK - Number of likes retrieved successfully
          content:
            application/json:
              schema:
                title: FetchLikesResponse
                type: object
                properties:
                  likeCount:
                    type: integer
        '404':
          description: Not Found - Post with the given ID does not exist

components:
  schemas:
    UserCredentials:
      title: UserCredentials
      type: object
      properties:
        hashedPassword: 
          type: string
        username: 
          type: string
        isAdmin:
          type: boolean
      required:
        - username
        - hashedPassword
        - isAdmin
    LikeRequest:
      title: LikeRequest
      type: object
      properties:
        username:
          type: string
        postId:
          type: integer
        status:
          type: boolean
      required:
        - username
        - postId
    FetchPostsResponse:
      title: FetchPostsResponse
      type: object
      properties:
        posts:
          type: array
          items:
            title: PostDTO
            type: object
            properties:
              id:
                type: integer
              categoryId:
                type: integer
              username:
                type: string
              title:
                type: string
              description:
                type: string
              address:
                type: string
              datePosted:
                type: string
                format: date-time
  responses:
    AuthResponse:
      description: OK
      content:
        application/json:
          schema:
            title: authResponse
            type: object
            properties:
              resp:
                type: integer
              authSuccess:
                type: boolean